+++ conditions/timer.go
@@ -242,7 +242,7 @@ func processTimerCondition(config *cfg.Config, block *hcl.Block, def *cfg.Condit
 	if t.hasInputExpr {
 		re, d := cfg.NewReactiveExpr(body.Input, config.EvalCtx(), func(ctx context.Context, v cty.Value) {
 			if v.Type() != cty.Bool || v.IsNull() || !v.IsKnown() {
-				config.Logger.Warn("condition input expression did not produce a boolean",
+				config.UserLogger.Warn("condition input expression did not produce a boolean",
 					zap.String("name", t.name), zap.String("type", v.Type().FriendlyName()))
 				return
 			}
@@ -255,7 +255,7 @@ func processTimerCondition(config *cfg.Config, block *hcl.Block, def *cfg.Condit
 	if cfg.IsExpressionProvided(body.Inhibit) {
 		re, d := cfg.NewReactiveExpr(body.Inhibit, config.EvalCtx(), func(ctx context.Context, v cty.Value) {
 			if v.Type() != cty.Bool || v.IsNull() || !v.IsKnown() {
-				config.Logger.Warn("condition inhibit expression did not produce a boolean",
+				config.UserLogger.Warn("condition inhibit expression did not produce a boolean",
 					zap.String("name", t.name), zap.String("type", v.Type().FriendlyName()))
 				return
 			}